Prepare_multilabel reproducible class_list (#83)
* Feat (prepare_multilabel): reproducible class_list

Added sort() to the code generating the class_list to ensure that it always is in the same order
This is crucial in applications that require a consistent class order
set() does not generate reproducible results by itself.

* Docs:  Ran tests and updated changelog

Ran nbdev_prepare and pytest with no problems. Added the fix made to prepare_multilabel in train_test_split to changelog