fix(tests): update to `glyles~=1.2.3a0`
Updating to 1.2.3a0 or later means pulling in a newer `pydot` — which
was needed to fix a dependency mismatch that was causing a test failure.

I'm updating all versions of `glyles`, not just the one under `[dev]`,
so that the tests and main application don't use different versions of
the same dependency. Otherwise test results could diverge from real
usage!